+++ app/models/subject.rb
@@ -15,6 +15,7 @@ class Subject < ActiveRecord::Base
   scope :with_entered_design_on_event, lambda {|design, event| where("subjects.id IN (select subject_events.subject_id from subject_events where subject_events.event_id = ? and subject_events.id IN (SELECT sheets.subject_event_id from sheets where sheets.deleted = ? and sheets.design_id = ? and sheets.subject_event_id IS NOT NULL))", event, false, design)}
   scope :with_unentered_design_on_event, lambda {|design, event| where("subjects.id IN (select subject_events.subject_id from subject_events where subject_events.event_id = ? and subject_events.id NOT IN (SELECT sheets.subject_event_id from sheets where sheets.deleted = ? and sheets.design_id = ? and sheets.subject_event_id IS NOT NULL))", event, false, design)}
   scope :without_design_on_event, lambda {|design, event| where("subjects.id NOT IN (select subject_events.subject_id from subject_events where subject_events.event_id = ? and subject_events.id IN (SELECT sheets.subject_event_id from sheets where sheets.deleted = ? and sheets.design_id = ?))", event, false, design)}
+  # scope :with_variable, lambda {|variable_id, value| where("subjects.id IN (select sheets.subject_id from sheets where sheets.deleted = ? and sheets.id IN (select sheet_variables.sheet_id from sheet_variables where variable_id = ? and response IN (?)))", false, variable_id, value)}
 
   # Model Validation
   validates_presence_of :project_id, :subject_code, :site_id
@@ -77,6 +78,10 @@ class Subject < ActiveRecord::Base
 
   def uploaded_files
     SheetVariable.where(sheet_id: self.sheets.select(:id)).includes(:variable).where(variables: { variable_type: 'file' }).order(created_at: :desc)
+  end
+
+  def has_value?(variable, value)
+    self.sheets.joins(:sheet_variables).where(sheet_variables: { variable_id: variable.id, response: value }).count >= 1
   end
 
 end