+++ db/migrate/20150722185820_add_variable_scope_to_randomization_scheme.rb
@@ -1,0 +1,6 @@
+class AddVariableScopeToRandomizationScheme < ActiveRecord::Migration
+  def change
+    add_column :randomization_schemes, :variable_id, :integer
+    add_column :randomization_schemes, :variable_value, :string
+  end
+end