+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - When site is used as a stratification factor, subjects can only be stratified to their own site
   - Site Editors can create and undo randomizations for subjects on their site
   - Site Viewers can view all existing randomization for subjects on their site
+  - Subject randomization criteria can be set to only allow distinguishing between eligible and ineligible subjects
   - **Permuted-block Algorithm**
     - Treatment Arms can now be specified along with a weight allocation
     - Different block sizes and allocations can now be specified